Update subject name replacements
Summary: SIMS has different subject abbreviations which need to be updated in dictSubjects. It would be preferable to update this so that the app shows the full name of the subjects that the student is taking. This helps avoid confusion about subjects such as "Pz" which is obscure.

Test Plan: Using multiple timetables, make sure that the vast majority of subjects are covered.

Differential Revision: https://phabricator.georgegarside.com/D1